Simpler class for string based scalars
