+++ src/Tgstation.Server.Host/GraphQL/Scalars/JwtType.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 
-using HotChocolate.Language;
-using HotChocolate.Types;
-
 namespace Tgstation.Server.Host.GraphQL.Scalars
 {
 	/// <summary>
-	/// A <see cref="ScalarType{TRuntimeType, TLiteral}"/> for encoded JSON Web Tokens.
+	/// A <see cref="StringScalarType"/> for encoded JSON Web Tokens.
 	/// </summary>
-	public sealed class JwtType : ScalarType<string, StringValueNode>
+	public sealed class JwtType : StringScalarType
 	{
 		/// <summary>
 		/// Initializes a new instance of the <see cref="JwtType"/> class.
@@ -19,20 +16,5 @@ namespace Tgstation.Server.Host.GraphQL.Scalars
 			Description = "Represents an encoded JSON Web Token";
 			SpecifiedBy = new Uri("https://datatracker.ietf.org/doc/html/rfc7519");
 		}
-
-		/// <inheritdoc />
-		public override IValueNode ParseResult(object? resultValue)
-			=> ParseValue(resultValue);
-
-		/// <inheritdoc />
-		protected override string ParseLiteral(StringValueNode valueSyntax)
-		{
-			ArgumentNullException.ThrowIfNull(valueSyntax);
-			return valueSyntax.Value;
-		}
-
-		/// <inheritdoc />
-		protected override StringValueNode ParseValue(string runtimeValue)
-			=> new(runtimeValue);
 	}
 }
